+++ coloraide/colors/_convert.py
@@ -46,7 +46,7 @@ class Convert:
     def _constrain_hue(cls, hue):
         """Constrain hue to 0 - 360."""
 
-        return hue % 360
+        return hue % 360 if not util.is_nan(hue) else hue
 
     @classmethod
     def _chromatic_adaption(cls, w1, w2, xyz):
@@ -110,6 +110,7 @@ class Convert:
         """Update from color."""
 
         if self is obj:
+            obj._coords = obj.null_adjust(obj._coords)
             return
 
         if not isinstance(obj, type(self)):
@@ -118,4 +119,5 @@ class Convert:
         for i, value in enumerate(obj.coords()):
             self._coords[i] = value
         self.alpha = obj.alpha
+        self._coords = self.null_adjust(self._coords)
         return self