+++ coloraide/colors/lch.py
@@ -10,7 +10,7 @@ from .. import util
 import re
 import math
 
-ACHROMATIC_THRESHOLD = 0.015
+ACHROMATIC_THRESHOLD = 0.02
 
 
 def lab_to_lch(lab):
@@ -18,17 +18,27 @@ def lab_to_lch(lab):
 
     l, a, b = lab
 
-    return (
-        l,
-        math.sqrt(math.pow(a, 2) + math.pow(b, 2)),
-        LCH._constrain_hue(math.atan2(b, a) * 180 / math.pi)
-    )
+    c = math.sqrt(math.pow(a, 2) + math.pow(b, 2))
+    h = math.atan2(b, a) * 180 / math.pi
+
+    # This is not actually part of the conversion, but is a fix-up
+    # for conversion getting a bit chaotic in regards to hue when
+    # chroma approaches zero. This fix-up is intended to make at
+    # least colors in the sRGB range a bit more stable with conversion
+    # and yield a hue of zero. This minimally affects the overall output.
+    # If a 100% accurate result is desired, then we'd want to avoid doing this.
+    if c < ACHROMATIC_THRESHOLD:
+        h = util.NaN
+
+    test = [l, c, LCH._constrain_hue(h)]
+    return test
 
 
 def lch_to_lab(lch):
     """LCH to LAB."""
 
     l, c, h = lch
+    h = util.no_nan(h)
 
     # If, for whatever reason (mainly direct user input),
     # if chroma is less than zero, clamp to zero.
@@ -85,12 +95,6 @@ class LCH(Cylindrical, Space):
         else:
             raise TypeError("Unexpected type '{}' received".format(type(color)))
 
-    def is_hue_null(self):
-        """Test if hue is null."""
-
-        l, c, h = self.coords()
-        return c < ACHROMATIC_THRESHOLD
-
     def hue_index(self):
         """Get hue index."""
 
@@ -133,6 +137,14 @@ class LCH(Cylindrical, Space):
         self._coords[2] = self._handle_input(value)
 
     @classmethod
+    def null_adjust(cls, coords):
+        """On color update."""
+
+        if coords[1] < ACHROMATIC_THRESHOLD:
+            coords[2] = util.NaN
+        return coords
+
+    @classmethod
     def translate_channel(cls, channel, value):
         """Translate channel string."""
 
@@ -144,11 +156,6 @@ class LCH(Cylindrical, Space):
             return parse.norm_alpha_channel(value)
         else:
             raise ValueError("Unexpected channel index of '{}'".format(channel))
-
-    def to_string(self, *, alpha=None, precision=util.DEF_PREC, fit=True, **kwargs):
-        """To string."""
-
-        return super().to_string(alpha=alpha, precision=precision, fit=fit)
 
     @classmethod
     def _to_lab(cls, lch):