+++ coloraide/css/colors/lch.py
@@ -29,16 +29,20 @@ class LCH(generic.LCH):
         super().__init__(color)
 
     def to_string(
-        self, *, alpha=None, precision=util.DEF_PREC, fit=True, **kwargs
+        self, *, alpha=None, precision=None, fit=True, **kwargs
     ):
         """Convert to CSS."""
+
+        if precision is None:
+            precision = self.parent.PRECISION
 
         options = kwargs
         if options.get("color"):
             return super().to_string(alpha=alpha, precision=precision, fit=fit, **kwargs)
 
-        alpha = alpha is not False and (alpha is True or self.alpha < 1.0)
-        coords = self.fit_coords() if fit else self.coords()
+        a = util.no_nan(self.alpha)
+        alpha = alpha is not False and (alpha is True or a < 1.0)
+        coords = util.no_nan(self.fit_coords() if fit else self.coords())
 
         if alpha:
             template = "lch({}%, {}, {}, {})" if options.get("comma") else "lch({}% {} {} / {})"
@@ -46,7 +50,7 @@ class LCH(generic.LCH):
                 util.fmt_float(coords[0], precision),
                 util.fmt_float(coords[1], precision),
                 util.fmt_float(coords[2], precision),
-                util.fmt_float(self.alpha, max(util.DEF_PREC, precision))
+                util.fmt_float(a, max(util.DEF_PREC, precision))
             )
         else:
             template = "lch({}%, {}, {})" if options.get("comma") else "lch({}% {} {})"
@@ -84,7 +88,7 @@ class LCH(generic.LCH):
                 channels.append(cls.translate_channel(-1, c))
         if len(channels) == 3:
             channels.append(1.0)
-        return channels
+        return cls.null_adjust(channels)
 
     @classmethod
     def match(cls, string, start=0, fullmatch=True):